refactor(signin page): code refactor
