--- app/routes/registration.tsx
@@ -1,3 +1,0 @@
-export default function Registration(){
-    return(<>WORK IN SUPER PROGRESSSSSSSS</>)
-}
